refactor: add support for sized collections
