+++ src/collections/mod.rs
@@ -1,5 +1,9 @@
 mod common;
 mod data;
+mod items;
 mod methods;
+mod migrate;
 pub use data::*;
+pub use items::*;
 pub use methods::*;
+pub use migrate::*;
