+++ src/process_subcommands.rs
@@ -4,8 +4,8 @@ use solana_client::{nonblocking::rpc_client::RpcClient as AsyncRpcClient, rpc_cl
 use crate::burn::burn_one;
 use crate::collections::{
     approve_delegate, check_collection_items, get_collection_items, migrate_collection,
-    revoke_delegate, set_and_verify_nft_collection, unverify_nft_collection, verify_nft_collection,
-    MigrateArgs,
+    revoke_delegate, set_and_verify_nft_collection, set_size, unverify_nft_collection,
+    verify_nft_collection, MigrateArgs,
 };
 use crate::decode::{decode_master_edition, decode_metadata};
 use crate::derive::{get_cmv2_pda, get_edition_pda, get_generic_pda, get_metadata_pda};
@@ -116,6 +116,12 @@ pub async fn process_collections(
             delegate_authority,
         } => revoke_delegate(client, keypair, collection_mint, delegate_authority),
 
+        CollectionsSubcommands::SetSize {
+            keypair,
+            collection_mint,
+            size,
+        } => set_size(client, keypair, collection_mint, size),
+
         CollectionsSubcommands::Migrate {
             keypair,
             mint_address,
@@ -143,6 +149,7 @@ pub async fn process_collections(
             method,
             api_key,
         } => get_collection_items(collection_mint, method, api_key).await,
+
         CollectionsSubcommands::CheckItems {
             collection_mint,
             item_list,
